% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ws.R
\name{get_ws}
\alias{get_ws}
\title{Load an Excel worksheet}
\usage{
get_ws(dir, patterns)
}
\arguments{
\item{dir}{Directory in which files are stored}

\item{patterns}{One or more regex patterns to match files.

Note that regex patterns are different from the glob (wildcard) patterns
that many users are accustomed to!}
}
\description{
This is just a wrapper function for XLConnect::readWorksheetFromFile()
for easier importing of Excel data.
}
\examples{
ws <- get_ws('data','GeoCenter 0524-1.xls')
}

