% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lcc.R
\name{graph_lcc}
\alias{graph_lcc}
\title{Largest connected component of a graph}
\usage{
graph_lcc(g)
}
\arguments{
\item{g}{igraph graph object.}
}
\description{
Many graphs, including social networks, include a "giant component",
meaning that most nodes in the network are connected to each other.
Many network metrics (e.g. betweenness centrality) are only meaningful
within a connected component, and connected graphs are easier to
visualize. This function returns the largest (weakly) connected
component of a graph.
}
\examples{
library(igraph)
g <- erdos.renyi.game(100,0.02)
lcc <- graph_lcc(g)
length(V(lcc))
}

