% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayors.R
\name{mayors}
\alias{mayors}
\title{Find "mayors" in a network}
\usage{
mayors(g, comm, n = 8, metric = page.rank, extra = NULL)
}
\arguments{
\item{g}{An igraph graph object.}

\item{comm}{An igraph communities object}

\item{n}{The number of mayors to find. By default, returns mayors for the
largest n communities.}

\item{metric}{Function to use in determining influence within a community}

\item{extra}{Additional communities whose mayors should be found.}
}
\description{
In a social network, community-detection algorithms can be used to find
cliques that are more strongly connected to each other than to the rest
of the network. For each community, the "mayor" is the node with the highest
influence. Finding mayors can often be a useful way of figuring out
whether the community has a definable theme or group identity.
}
\examples{
library(dplyr)
library(igraph)
g <- get_ws('data','GeoCenter 0524-1.xls') \%>\%
    ws_to_graph() \%>\%
    graph_lcc()
fg <- g \%>\%
    as.undirected() \%>\%
    simplify() \%>\%
    fastgreedy.community()
mayors(g,fg)
}

