% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_bar.R
\name{sma_bar}
\alias{sma_bar}
\title{Bar chart showing network influencers}
\usage{
sma_bar(g, n = 8, highlight = NULL, extra = NULL)
}
\arguments{
\item{g}{An igraph graph object.}

\item{n}{Number of influencers to display. The maximum value (8) is set by
the number of colors available from RColorBrewer.}

\item{highlight}{The results of a scoring function according to which
highlighted nodes should be chosen. Calls igraph::page.rank() if nothing
is passed.  Calculating the scoring in advance lets you economize if you'll
be using the same scoring function for an sma_plot() call and an sma_bar()
call.}

\item{extra}{A list of node names to be highlighted in addition to the
top-ranked ones.}
}
\description{
This function is used together with sma_plot(), and gives a color-coded
guide to the top influencers and their relative influence scores.
}
\examples{
library(dplyr)
sample_tweets \%>\%
    ws_to_graph() \%>\%
    sma_bar()
}

