% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_plot.R
\name{sma_plot}
\alias{sma_plot}
\title{Draw a network, highlighting key influencers}
\usage{
sma_plot(g, n = 8, highlight = NULL, layout = NULL, extra = NULL)
}
\arguments{
\item{g}{An igraph graph object.}

\item{n}{Number of nodes to highlight. The maximum value (8) is set by
the number of colors available from RColorBrewer.}

\item{highlight}{The results of a scoring function according to which
highlighted nodes should be chosen. Calls igraph::page.rank() if nothing
is passed. Calculating the scoring in advance lets you economize if you'll
be using the same scoring function for an sma_plot() call and an sma_bar()
call.}

\item{layout}{Results of an igraph layout routine
(layout.fruchterman.reingold is called if nothing is passed). Because these
calculations can be time consuming, it makes sense to obtain the layout in
advance if you'll be making several versions of the same plot.}

\item{extra}{A list of node names to be highlighted in addition to the
top-ranked ones.}
}
\description{
This function draws a (potentially large) social network. A metric function
(by default igraph::page.rank) is used to identify up to eight top
influencers, and these are highlighted by edge coloring.
}
\examples{
library(dplyr)
get_ws('data','GeoCenter 0524-1.xls') \%>\%
    ws_to_graph() \%>\%
    graph_lcc() \%>\%
    sma_plot()
}

