% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_to_graph.R
\name{ws_to_graph}
\alias{ws_to_graph}
\title{Convert a worksheet from Crimson Hexagon to an igraph object}
\usage{
ws_to_graph(ws)
}
\arguments{
\item{ws}{Data frame returned, for example, by get_ws()}
}
\description{
This function converts Excel worksheet containing a Crimson Hexagon data
export to a directed graph. Nodes are named after Twitter handles (including
the '@' prefix) and edges point from a retweeter to a retweetee.
}
\details{
TODO: it might be handy to include the ability to build a graph from
mentions as well, rather than just retweets.
}
\examples{
library(dplyr)
g <- sample_tweets \%>\%
    ws_to_graph()

}
\author{
Craig Jolley, cjolley.usaid@gmail.com
}

